HotkeySettings: Replaced pykdeuic4 generated code with cleaned up original ui file. Moved HotkeySettings class from dialogs.py to hotkeysettings.py.
