+++ lib/autokey/qtui/hotkeysettings.py
@@ -1,88 +1,41 @@
-#!/usr/bin/env python
-# coding=UTF-8
-#
-# Generated by pykdeuic4 from hotkeysettings.ui on Sun Mar  4 11:39:40 2012
-#
-# WARNING! All changes to this file will be lost.
-from PyKDE4 import kdecore
-from PyKDE4 import kdeui
-from PyQt4 import QtCore, QtGui
+# Copyright (C) 2011 Chris Dekter
+# Copyright (C) 2018 Thomas Hess <thomas.hess@udo.edu>
 
-try:
-    _fromUtf8 = QtCore.QString.fromUtf8
-except AttributeError:
-    _fromUtf8 = lambda s: s
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
 
-class Ui_Form(object):
-    def setupUi(self, Form):
-        Form.setObjectName(_fromUtf8("Form"))
-        Form.resize(400, 300)
-        self.verticalLayout = QtGui.QVBoxLayout(Form)
-        self.verticalLayout.setObjectName(_fromUtf8("verticalLayout"))
-        self.modsLabel = QtGui.QLabel(Form)
-        self.modsLabel.setObjectName(_fromUtf8("modsLabel"))
-        self.verticalLayout.addWidget(self.modsLabel)
-        self.horizontalLayout = QtGui.QHBoxLayout()
-        self.horizontalLayout.setContentsMargins(-1, -1, -1, 10)
-        self.horizontalLayout.setObjectName(_fromUtf8("horizontalLayout"))
-        self.controlButton = KPushButton(Form)
-        self.controlButton.setCheckable(True)
-        self.controlButton.setChecked(False)
-        self.controlButton.setObjectName(_fromUtf8("controlButton"))
-        self.horizontalLayout.addWidget(self.controlButton)
-        self.altButton = KPushButton(Form)
-        self.altButton.setCheckable(True)
-        self.altButton.setChecked(False)
-        self.altButton.setObjectName(_fromUtf8("altButton"))
-        self.horizontalLayout.addWidget(self.altButton)
-        self.shiftButton = KPushButton(Form)
-        self.shiftButton.setCheckable(True)
-        self.shiftButton.setChecked(False)
-        self.shiftButton.setObjectName(_fromUtf8("shiftButton"))
-        self.horizontalLayout.addWidget(self.shiftButton)
-        self.superButton = KPushButton(Form)
-        self.superButton.setCheckable(True)
-        self.superButton.setChecked(False)
-        self.superButton.setObjectName(_fromUtf8("superButton"))
-        self.horizontalLayout.addWidget(self.superButton)
-        self.hyperButton = KPushButton(Form)
-        self.hyperButton.setCheckable(True)
-        self.hyperButton.setChecked(False)
-        self.hyperButton.setObjectName(_fromUtf8("hyperButton"))
-        self.horizontalLayout.addWidget(self.hyperButton)
-        self.metaButton = KPushButton(Form)
-        self.metaButton.setCheckable(True)
-        self.metaButton.setChecked(False)
-        self.metaButton.setObjectName(_fromUtf8("metaButton"))
-        self.horizontalLayout.addWidget(self.metaButton)
-        self.verticalLayout.addLayout(self.horizontalLayout)
-        self.horizontalLayout_2 = QtGui.QHBoxLayout()
-        self.horizontalLayout_2.setContentsMargins(-1, -1, -1, 5)
-        self.horizontalLayout_2.setObjectName(_fromUtf8("horizontalLayout_2"))
-        self.keyLabel = QtGui.QLabel(Form)
-        self.keyLabel.setObjectName(_fromUtf8("keyLabel"))
-        self.horizontalLayout_2.addWidget(self.keyLabel)
-        self.setButton = QtGui.QPushButton(Form)
-        self.setButton.setObjectName(_fromUtf8("setButton"))
-        self.horizontalLayout_2.addWidget(self.setButton)
-        self.verticalLayout.addLayout(self.horizontalLayout_2)
-        self.kseparator = KSeparator(Form)
-        self.kseparator.setObjectName(_fromUtf8("kseparator"))
-        self.verticalLayout.addWidget(self.kseparator)
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+# GNU General Public License for more details.
 
-        self.retranslateUi(Form)
-        QtCore.QMetaObject.connectSlotsByName(Form)
+# You should have received a copy of the GNU General Public License
+# along with this program. If not, see <http://www.gnu.org/licenses/>.
 
-    def retranslateUi(self, Form):
-        Form.setWindowTitle(kdecore.i18n(_fromUtf8("Form")))
-        self.modsLabel.setText(kdecore.i18n(_fromUtf8("Modifiers:")))
-        self.controlButton.setText(kdecore.i18n(_fromUtf8("Control")))
-        self.altButton.setText(kdecore.i18n(_fromUtf8("Alt")))
-        self.shiftButton.setText(kdecore.i18n(_fromUtf8("Shift")))
-        self.superButton.setText(kdecore.i18n(_fromUtf8("Super")))
-        self.hyperButton.setText(kdecore.i18n(_fromUtf8("Hyper")))
-        self.metaButton.setText(kdecore.i18n(_fromUtf8("Meta")))
-        self.keyLabel.setText(kdecore.i18n(_fromUtf8("Key: %s")))
-        self.setButton.setText(kdecore.i18n(_fromUtf8("Press to set")))
 
-from PyKDE4.kdeui import KSeparator, KPushButton
+from PyQt4 import uic
+
+from .common import get_ui_qfile
+
+from .. import iomediator
+
+ui_file = get_ui_qfile("hotkeysettings")
+HotkeySettingsBase = uic.loadUiType(ui_file)
+ui_file.close()
+
+
+class HotkeySettings(*HotkeySettingsBase):
+
+    def __init__(self, parent):
+        super(HotkeySettings, self).__init__(parent)
+        self.setupUi(self)
+
+    # ---- Signal handlers
+
+    def on_setButton_pressed(self):
+        self.setButton.setEnabled(False)
+        self.keyLabel.setText("Press a key or combination...")  # TODO: i18n
+        self.grabber = iomediator.KeyGrabber(self.parentWidget())
+        self.grabber.start()
